PackageGenerator.jl

PackageGenerator.Package

PackageGenerator.configure

PackageGenerator.generate

PackageGenerator.update_configuration

PackageGenerator.Package — Type.

Package(package_name)

Generate a default package based on a package name. To customize your package, you can edit any of its fields before generateing it. The package will include all the fields discussed in configure, package_name, as well as the following defaults (override with keyword arguments).

repo_name = package_name * ".jl"

The name of your repository on github.

year = Dates.today() |> Dates.year

The copyright year. Defaults to the current year.

source

PackageGenerator.configure — Method.

configure(github_token, appveyor_token)

Write out a user configuration file (".package_generator.json" in your package directory).

Requires a github_token, which can be generated here. Make sure to check the "public_repo" and "delete_repo" scopes (if you want to be able to delete half-created repositories in case PackageGenerator hits an error).

Also requires an appveyor_token, which can be found here

Includes various defaults that can be overwritten with keyword arguments.

ssh_keygen_file = "ssh-keygen"

The path to ssh-keygen. ssh-keygen makes a pair of keys that allows Travis to communicate with Github. For Linux users with git installed, the default file should be fine. For Windows users with git installed, try ssh_keygen_file = "C:/Program Files/Git/usr/bin/ssh-keygen".

travis_token = get_travis_token(github_token)

A travis token can be generated automatically from your github token

package_directory = Pkg.Dir.path()

Where to put your new packages

authors = LibGit2.getconfig("user.name", "YOUR_NAME")

Who are the authors of your packages?

user_name = LibGit2.getconfig("github.user", "YOUR_GITHUB_USER_NAME")

What is your github username?

sync_time = 60

How many seconds to wait for API calls to complete.

source

PackageGenerator.generate — Method.

generate(package::Package)

Generate a Package with some nice bells and whistles. These include:

a matching github repository

testing with travis

testing with appveyor

generated documentation that

automatically syncs to changes on github

includes doctests as part of your package testing suite

Of course, this means you need github, travis, and appveyor accounts.

If you haven't set up an ssh key for git, follow the instructions here.

The license will be the MIT license. To change, use PkgDev.Generate.license

source

PackageGenerator.update_configuration — Method.

update_configuration(; kwargs...)

Update any of the user configurations discussed in configure.

source

If you would like to run tests locally, first configure then try running tests/local_tests.jl (and report any errors!). This will test parts of the package that are not tested through CI.