PackageGenerator.jl

PackageGenerator.generate

PackageGenerator.generate_offline

PackageGenerator.generate_online

PackageGenerator.generate — Method.

generate(package;
    license = "MIT",
    ssh_keygen_file = "ssh-keygen",
    appveyor_token = "")

Generate a package named package with some nice bells and whistles. These include:

a matching github repository

an activated repository on travis

generated documentation that

automatically syncs to changes on github

includes doctests as part of your package testing suite

Of course, this means you need both a github and a travis account.

The package defaults to the "MIT" license. See PkgDev for other options.

ssh-keygen makes a pair of keys that allows Travis to communicate with Github. For Linux users with git installed, the default file should be fine. For Windows users with git installed, try ssh_keygen_file = "C:/Program Files/Git/usr/bin/ssh-keygen".

Your appveyor_token is available here. Include the token in order to automatically turn on the appveyor for your repo.

For LibGit2.push to work, follow the instructions here

By default, tests will fail. Documentation will not build until tests pass.

source

PackageGenerator.generate_offline — Method.

generate_offline(package, license = "MIT")

Only the offline components of generate (see for more documentation)

Create a local git repository containing package files.

source

PackageGenerator.generate_online — Method.

generate_online(repo_name; ssh_keygen_file = "ssh-keygen", appveyor_token = "")

Only the online components of generate (see for more documentation)

Set up a linked github and travis account for a certain repository.

source

Advanced usage

For advanced users, there are several undocumented keywords for generate. They are listed below, along with their default values.

path = joinpath(Pkg.Dir.path(), package)

The folder where files will be generated

authors = LibGit2.getconfig("user.name", "")

The authors of the package. Interpolated directly into the license.

years = PkgDev.Generate.copyright_year()

The copyright year. Interpolated directly into the license.

user = PkgDev.GitHub.user()

Your github username

repo_name = string(package, ".jl")

The desired name of the github repository

github_token = PkgDev.GitHub.token()

Your token to access the GitHub API

travis_token = get_travis_token(github_token)

Your token to access the Travis API

appveyor_slug = default_appveyor_slug(repo_name)

The name of the appveyor project for your repository. If an appveyor token is given, the actual appveyor_slug will be used instead.