PackageGenerator.jl

PackageGenerator.Package

PackageGenerator.configure

PackageGenerator.generate

PackageGenerator.update_configuration

PackageGenerator.Package — Type.

Package(package_name)

Generate a default package based on a package name. To customize you pacakge, you can edit any of its fields before generateing it. The package will include all the fields discussed in configure, package_name, as well as:

repo_name = string(package_name, ".jl")

The name of your repository on github.

year = string(Dates.year(Dates.today()))

The copyright year. Defaults to the current year.

source

PackageGenerator.configure — Method.

configure(github_token, appveyor_token)

Write out a user configuration file.

Requires a github_token, which can be generated here. Make sure to check the "public_repo" and "delete_repo" scopes (if you want to be able to delete half-created repositories in case PackageGenerator hits an error).

Also requires an appveyor_token, which can be found here

Includes various defaults that can be overwritten with keyword arguments.

ssh_keygen_file = "ssh-keygen"

The path to ssh-keygen. ssh-keygen makes a pair of keys that allows Travis to communicate with Github. For Linux users with git installed, the default file should be fine. For Windows users with git installed, try ssh_keygen_file = "C:/Program Files/Git/usr/bin/ssh-keygen".

travis_token = get_travis_token(github_token)

A travis token can be generated automatically from your github token

package_directory = Pkg.Dir.path()

Where to put your new packages

license = "MIT"

What license to use. See PkgDev.jl for more options.

authors = LibGit2.getconfig("user.name", "")

Who are the authors of your packages?

user_name = LibGit2.getconfig("github.user", "")

What is your github username?

sync_time = 60

How many seconds to wait for API calls to complete. 60 is overly cautious by design.

source

PackageGenerator.generate — Method.

generate(package::Package)

Generate a package named package with some nice bells and whistles. These include:

a matching github repository

an activated repository on travis (and optionally, an activated appveyor project)

generated documentation that

automatically syncs to changes on github

includes doctests as part of your package testing suite

Of course, this means you need both a github and a travis account. If you haven't set up an ssh key for git, follow the instructions here.

You must include a Package

source

PackageGenerator.update_configuration — Method.

update_configuration(; kwargs)

Update any of the user configurations discussed in configure.

source

Advanced usage

For advanced users, there are several undocumented keywords for generate. They are listed below, along with their default values.

path = joinpath(Pkg.Dir.path(), package)

The folder where files will be generated

authors = LibGit2.getconfig("user.name", "")

The authors of the package. Interpolated directly into the license.

years = PkgDev.Generate.copyright_year()

The copyright year. Interpolated directly into the license.

user = PkgDev.GitHub.user()

Your github username

repo_name = string(package, ".jl")

The desired name of the github repository

github_token = PkgDev.GitHub.token()

Your token to access the GitHub API. The default token does not have permissions to delete repositories, which will cause a hiccup in the case of an error. Make your own with permission to delete repositories here

travis_token = get_travis_token(github_token)

Your token to access the Travis API

appveyor_slug = default_appveyor_slug(repo_name)

The name of the appveyor project for your repository. If an appveyor token is given, the actual appveyor_slug will be used instead.