PackageGenerator.jl

Documentation for PackageGenerator.jl

PackageGenerator.generate

PackageGenerator.generate — Method.

generate(package; license = "MIT", path_to_ssh_keygen = "")

Generate a package named package with some nice bells and whistles. These include:

a matching github repository

an activated repository on travis

generated documentation that

automatically syncs to changes on github

includes doctests as part of your package testing suite

The package defaults to the "MIT" license. See PkgDev for other options.

ssh-keygen is required. If it is not on your path, make sure to include a path to it. ssh-keygen is sometimes packaged with git. If you are on Windows, PackageGenerator will automatically check "C:/Program Files/Git/usr/bin" for it. If you are on Ubuntu, and don't yet have it installed, try sudo apt-get install openssh-client. Then you won't need to include a path.

source